requirements.txt
# Web framework
streamlit>=1.32.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Sentiment analysis
vaderSentiment>=3.3.2
transformers>=4.36.0
torch>=2.1.0
scipy>=1.11.0

# Data collection / scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
cloudscraper>=1.2.71
twikit==1.3.0

# Visualization
plotly>=5.18.0
wordcloud>=1.9.0
matplotlib>=3.8.0

# Utilities
python-dotenv>=1.0.0
